+++ DataAccess/EntityModelFundabien/entities/Pais.cs
@@ -8,7 +8,7 @@ namespace EntityModelFundabien.entities
     public class Pais
     {
         [Key]
-        public double idPais { get; set; }
+        public Int64 idPais { get; set; }
         public string nombre { get; set; }
         public List<Departamento> departamentos { get; set; }
     }
