+++ DataAccess/EntityModelFundabien/entities/PersonaEncargada.cs
@@ -8,9 +8,9 @@ namespace EntityModelFundabien.entities
     public class PersonaEncargada
     {
         [Key]
-        public double idPersonaEncargada { get; set; }
-        public double idPaciente { get; set; }
-        public double idPersona { get; set; }
+        public Int64 idPersonaEncargada { get; set; }
+        public Int64 idPaciente { get; set; }
+        public Int64 idPersona { get; set; }
         public string parentezco { get; set; }
         public bool estaActivo { get; set; }
