+++ DataAccess/EntityModelFundabien/entities/Telefono.cs
@@ -8,8 +8,8 @@ namespace EntityModelFundabien.entities
     public class Telefono
     {
         [Key]
-        public double idTelefono { get; set; }
-        public double idPersona { get; set; }
+        public Int64 idTelefono { get; set; }
+        public Int64 idPersona { get; set; }
         public string telefono { get; set; }
         public DateTime fechaRegistro { get; set; }
         public bool estaHabilitado { get; set; }
